fix Alchemistry goh & roots
